all: move api into its own directory
