+++ api/modules/emoji.js
