+++ api/modules/stream/manage.js
