+++ api/modules/util/currentCommit.js
