+++ api/modules/util/misc.js
